Add bare_sync_track fixture
This bare_* sort of fixture is cleaner for testing methods. This way,
all of the properties referenced by the method must be set
manually by the test case, so there will be no pollution by default
values. The only purpose is to obtain a reference to the method needing
to be tested.