Calculate NotesPerSecond over actual input interval
Before this, it was defining the interval over which NPS should be
calculated by the first NoteEvent after/on the beginning of the interval
and the last NoteEvent before/on the end of the interval. This was
unintuitive, so now we are instead actually treating the input values
as ends of the interval.